WIP: Overhaul sound system (first attempt)
See 78e7d1a748acdebd1c1acdd6c8e34cb7f68559f5, the redesign of this commit.

This branch was my unfinished first attempt at redesigning the sound subsystem
for planets. It was "too dependent" on the signal crate and its strict typing,
for which I cannot resolve at compile time (namely, I don't know how many
channels the sink will have until I instantiate it). It relied on 'static
lifetimes a lot, which would break if in the future I extended the asset module
to potentially load assets from files, or the network. It used a lot of trait
objects, which is less clear and potentially slower than the match statement
backing Source structs in the version on master. There were more layers of
indirection (even though I did still have to box Sources on the final version,
so I didn't have a recursive struct). The new version has some sharp edges
(such as the custom panic hook) but not nearly as many as this.

I knew it was back to the drawing board when I finished writing a match
statement with four cases, all identical :P